Story: Added exercise picker.
This is the basic feature of the app:
adding exercises from the list to
our current workout.

I do this by creating a new component,
`ExercisePicker`, and adding state
to the root `App` component.

I then coordinate state changes by
creating the `appendToWorkout` function,
which is passed to `ExercisePicker` to enable
selections to affect the app state.